+++ src/rinoh/flowable.py
@@ -831,6 +831,8 @@ class GroupedLabeledFlowables(GroupedFlowables):
 class FloatStyle(FlowableStyle):
     float = Attribute(Bool, False, 'Float the flowable to the top or bottom '
                                    'of the page')
+    sideways = Attribute(Bool, False, 'Render the flowable onto a separate '
+                                      'page, rotated 90 degrees')
 
 
 class Float(Flowable):
@@ -847,6 +849,10 @@ class Float(Flowable):
     """
 
     def flow(self, container, last_descender, state=None, **kwargs):
+        if (self.get_style('sideways', container)
+                and self.get_id(container.document) not in container.document.registered_sideways_floats):
+            container.document.add_sideways_float(self)
+            return 0, 0, last_descender
         if self.get_style('float', container):
             id = self.get_id(container.document)
             if id not in container.document.floats: